\name{ccChooser-package}
\alias{ccChooser-package}
\alias{ccChooser}
\docType{package}
\title{
Developing a core collections
}
\description{
ccChooser can be used to developing and evaluation of core collections for germplasm collections (entire collection). 
This package used to develop a core collection for biological resources like genbanks. 
A core collection is defined as a sample of accessions that represent, with the lowest possible level of redundancy, the genetic diversity 
(the richness of gene or genotype categories) of the entire collection. 
The establishing a core collection that represents genetic diversity of the entire collection with minimum loss of its original diversity 
and minimum redundancies is an important problem for gene-banks curators and crop breeders. 
ccChooser establish core collection base on phenotypic data (agronomic, morphological, phenological).
}
\details{
\tabular{ll}{
Package: \tab ccChooser\cr
Type: \tab Package\cr
Version: \tab 0.2.6\cr
Date: \tab 2012-09-25\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Marcin Studnicki and Konrad Debski
Maintainer: Marcin Studnicki <marcin_studnicki@sggw.pl>
}
