\name{dactylis_CC}
\alias{dactylis_CC}
\docType{data}
\title{
Core collection of orchadgrass
}
\description{
Core collection of orchadgrass (Dactylis glomerata L.) obtain from enitre collection, held at Botanical Garden of Plant Breeding and Acclimatization 
Institute in Bydgoszcz, Poland. Core collection establish used stratified method with Pro allocation method and random sampling. 
}
\usage{data(dactylis_CC)}
\format{
  A data frame with 197 accessions on the following 8 variables.
  \describe{
    \item{\code{X1}}{plant height (cm)}
    \item{\code{X2}}{inflorescence length }
    \item{\code{X3}}{leaf width (cm)}
    \item{\code{X4}}{leaf length (cm)}
    \item{\code{X5}}{total seasonal yield (kg)}
    \item{\code{X6}}{1000 kernel weight (g)}
    \item{\code{X7}}{number of days since the 1st April to inflorescence emergence}
    \item{\code{X8}}{number of days since the 1st April to inflorescence}
  }
}

\examples{
data(dactylis_CC)
data(dactylis_CC)
}
\keyword{datasets}
